added pay system

+++ src/main/java/com/alisa/moviereservationsystem/dto/returnDto/SeatReturnDto.java
@@ -3,5 +3,7 @@ package com.alisa.moviereservationsystem.dto.returnDto;
 import com.alisa.moviereservationsystem.models.enums.SeatStatus;
 import com.alisa.moviereservationsystem.models.enums.SeatType;
 
-public record SeatReturnDto(Long id, Integer seatNumber, Integer rowNumber, Float price, SeatType type, SeatStatus status, Long hallId) {
+import java.math.BigDecimal;
+
+public record SeatReturnDto(Long id, Integer seatNumber, Integer rowNumber, BigDecimal price, SeatType type, SeatStatus status, Long hallId) {
 }
